Portfolio: Trailing slash on void elements has no effect
Consider the following two cases for marking up an img element for the image at http://www.example.com/logo.svg:

img element without trailing slash:

<img alt=SVG src=http://www.example.com/logo.svg>
img element with trailing slash:

<img alt=SVG src=http://www.example.com/logo.svg/>
In the first case (without the trailing slash), http://www.example.com/logo.svg becomes the value of the src attribute in the DOM, as expected.

However, in the second case (with the trailing slash), http://www.example.com/logo.svg/ unexpectedly becomes the value of the src attribute in the DOM — which breaks display of the image.